Optimize findSessionsByShop
For each session added to the redis store, we add a key based on the
shop name and add the session id key to a set of all session ids for that
shop. This allows us to quickly find all sessions for a given shop, and
avoids having to iterate over all sessions in the store to find the ones
for a given shop.