[Already reviewed] Remove node-fetch dependency and rely on globalThis.fetch instead
This affects these packages the @shopify/shopify-api package

Change toMatchObject instead of toStrictEqual in billing tests. Previously the object that was returned was returned created by node-fetch, meaning it passed the strict check.  Now it's created from gloablThis.Response, so it does not match the strict equslity check.  I've confirmed the the JSON content is identical though, so changing to toMatchObject is safe.