API: Update the E2E test setup for Node
Node 20 has issues using fetch for localhost requests which the now removed node-fetch package did not have.  The solution is to fallback to http methods provided by Node if the E2E test enviroment is Node.  Production apps should not be affected by this because our E2E setup is making requests to localhost in place of Shopify.