Change versioning-strategy to increase
I misunderstood how increase-if-necessary works.  How it works:

1. Will only update the package.json if the range doesnt match
2. If the range matches, it will update the package-lock.yml file

I chose increase-if-necessary because I thought it would reduce PR's. It doesn't.  We get the same number of PR's, just some of them don't touc hthe package.json file.

On top of that, increase-if-necessary seems problematic since it may change the package-lock.yml file, but not the actual package.json file. This means our local version of a package could drift from the released version of a package, even with the same version number.  That seems like it would be hard to debug.